first working version

+++ PrecisionJump/PrecisionJump/Program.cs
@@ -11,10 +11,14 @@ namespace PrecisionJump
     class Program
     {
         const string ADBPATH = @"C:\Program Files (x86)\Android\android-sdk\platform-tools\adb.exe";
-        const byte TargetWR = 245, TargetWG = 245, TargetWB = 245;
-        const byte PersonR = 49, PersonG = 47, PersonB = 78, PersonTolerance = 20;
+        const byte TargetWR = 245, TargetWG = 245, TargetWB = 245, TargetWMaxWidthMin = 35;
+        const byte PersonR = 49, PersonG = 47, PersonB = 78, PersonTolerance = 18;
         const int PersonSearchWith = 300, PersonSearchHeight = 300;
-        const double InverseSpeed = 1.4;
+        const double InvSpeedRight = 202.82343, OffsetRight = 41.5, InvSpeedLeft = 203.84435, OffsetLeft = 40.5;
+        const double Inch2Cm = 2.54, dpiX = 391.885, dpiY = 381.0;
+
+        // adb shell dumpsys display
+        // 1080 x 1920, 60.0 fps, density 3.0, 391.885 x 381.0 dpi
 
         static string CmdAdb(string arguments)
         {
@@ -35,172 +39,6 @@ namespace PrecisionJump
             return ret;
         }
 
-        static unsafe void DetectColorWithUnsafe(Bitmap image, byte searchedR, byte searchedG, int searchedB, int tolerance)
-        {
-            BitmapData imageData = image.LockBits(new Rectangle(0, 0, image.Width,
-              image.Height), ImageLockMode.ReadWrite, PixelFormat.Format24bppRgb);
-            int bytesPerPixel = 3;
-
-            byte* scan0 = (byte*)imageData.Scan0.ToPointer();
-            int stride = imageData.Stride;
-
-            byte unmatchingValue = 0;
-            byte matchingValue = 255;
-            int toleranceSquared = tolerance * tolerance;
-
-            for (int y = 0; y < imageData.Height; y++)
-            {
-                byte* row = scan0 + (y * stride);
-
-                for (int x = 0; x < imageData.Width; x++)
-                {
-                    // Watch out for actual order (BGR)!
-                    int bIndex = x * bytesPerPixel;
-                    int gIndex = bIndex + 1;
-                    int rIndex = bIndex + 2;
-
-                    byte pixelR = row[rIndex];
-                    byte pixelG = row[gIndex];
-                    byte pixelB = row[bIndex];
-
-                    int diffR = pixelR - searchedR;
-                    int diffG = pixelG - searchedG;
-                    int diffB = pixelB - searchedB;
-
-                    int distance = diffR * diffR + diffG * diffG + diffB * diffB;
-
-                    row[rIndex] = row[bIndex] = row[gIndex] = distance >
-                      toleranceSquared ? unmatchingValue : matchingValue;
-                }
-            }
-
-            image.UnlockBits(imageData);
-        }
-
-        //static void DetectColorWithMarshal(Bitmap image, byte searchedR, byte searchedG, int searchedB, int tolerance)
-        //{
-        //    BitmapData imageData = image.LockBits(new Rectangle(0, 0, image.Width,
-        //      image.Height), ImageLockMode.ReadWrite, PixelFormat.Format24bppRgb);
-
-        //    byte[] imageBytes = new byte[Math.Abs(imageData.Stride) * image.Height];
-        //    IntPtr scan0 = imageData.Scan0;
-
-        //    Marshal.Copy(scan0, imageBytes, 0, imageBytes.Length);
-
-        //    byte unmatchingValue = 0;
-        //    byte matchingValue = 255;
-        //    int toleranceSquared = tolerance * tolerance;
-
-        //    for (int i = 0; i < imageBytes.Length; i += 3)
-        //    {
-        //        byte pixelB = imageBytes[i];
-        //        byte pixelR = imageBytes[i + 2];
-        //        byte pixelG = imageBytes[i + 1];
-
-        //        int diffR = pixelR - searchedR;
-        //        int diffG = pixelG - searchedG;
-        //        int diffB = pixelB - searchedB;
-
-        //        int distance = diffR * diffR + diffG * diffG + diffB * diffB;
-
-        //        imageBytes[i] = imageBytes[i + 1] = imageBytes[i + 2] = distance >
-        //          toleranceSquared ? unmatchingValue : matchingValue;
-        //    }
-
-        //    Marshal.Copy(imageBytes, 0, scan0, imageBytes.Length);
-
-        //    image.UnlockBits(imageData);
-        //}
-
-        //static unsafe void DetectColorWithUnsafeParallel(Bitmap image, byte searchedR, byte searchedG, int searchedB, int tolerance)
-        //{
-        //    BitmapData imageData = image.LockBits(new Rectangle(0, 0, image.Width,
-        //      image.Height), ImageLockMode.ReadWrite, PixelFormat.Format24bppRgb);
-        //    int bytesPerPixel = 3;
-
-        //    byte* scan0 = (byte*)imageData.Scan0.ToPointer();
-        //    int stride = imageData.Stride;
-
-        //    byte unmatchingValue = 0;
-        //    byte matchingValue = 255;
-        //    int toleranceSquared = tolerance * tolerance;
-
-        //    Task[] tasks = new Task[4];
-        //    for (int i = 0; i < tasks.Length; i++)
-        //    {
-        //        int ii = i;
-        //        tasks[i] = Task.Factory.StartNew(() =>
-        //        {
-        //            int minY = ii < 2 ? 0 : imageData.Height / 2;
-        //            int maxY = ii < 2 ? imageData.Height / 2 : imageData.Height;
-
-        //            int minX = ii % 2 == 0 ? 0 : imageData.Width / 2;
-        //            int maxX = ii % 2 == 0 ? imageData.Width / 2 : imageData.Width;
-
-        //            for (int y = minY; y < maxY; y++)
-        //            {
-        //                byte* row = scan0 + (y * stride);
-
-        //                for (int x = minX; x < maxX; x++)
-        //                {
-        //                    int bIndex = x * bytesPerPixel;
-        //                    int gIndex = bIndex + 1;
-        //                    int rIndex = bIndex + 2;
-
-        //                    byte pixelR = row[rIndex];
-        //                    byte pixelG = row[gIndex];
-        //                    byte pixelB = row[bIndex];
-
-        //                    int diffR = pixelR - searchedR;
-        //                    int diffG = pixelG - searchedG;
-        //                    int diffB = pixelB - searchedB;
-
-        //                    int distance = diffR * diffR + diffG * diffG + diffB * diffB;
-
-        //                    row[rIndex] = row[bIndex] = row[gIndex] = distance >
-        //                        toleranceSquared ? unmatchingValue : matchingValue;
-        //                }
-        //            }
-        //        });
-        //    }
-
-        //    Task.WaitAll(tasks);
-
-        //    image.UnlockBits(imageData);
-        //}
-
-        //static unsafe void DetectColorWithUnsafe(Bitmap image, byte searchedR, byte searchedG, int searchedB)
-        //{
-        //    BitmapData imageData = image.LockBits(new Rectangle(0, 0, image.Width,
-        //      image.Height), ImageLockMode.ReadWrite, PixelFormat.Format24bppRgb);
-        //    int bytesPerPixel = 3;
-
-        //    byte* scan0 = (byte*)imageData.Scan0.ToPointer();
-        //    int stride = imageData.Stride;
-
-        //    byte unmatchingValue = 0;
-        //    byte matchingValue = 255;
-
-        //    for (int y = 0; y < imageData.Height; y++)
-        //    {
-        //        byte* row = scan0 + (y * stride);
-
-        //        for (int x = 0; x < imageData.Width; x++)
-        //        {
-        //            // Watch out for actual order (BGR)!
-        //            int bIndex = x * bytesPerPixel;
-        //            int gIndex = bIndex + 1;
-        //            int rIndex = bIndex + 2;
-        //            row[rIndex] = row[bIndex] = row[gIndex] =
-        //                ((row[rIndex] != searchedR)
-        //                || (row[gIndex] != searchedG)
-        //                || (row[bIndex] != searchedB)) ? unmatchingValue : matchingValue;
-        //        }
-        //    }
-
-        //    image.UnlockBits(imageData);
-        //}
-
         static unsafe void FindTargetWWithUnsafe(Bitmap image, byte searchedR, byte searchedG, int searchedB, out float centerX, out float centerY, out int count)
         {
             centerX = centerY = 0;
@@ -218,31 +56,49 @@ namespace PrecisionJump
             for (int y = ymin; y < ymax; y++)
             {
                 byte* row = scan0 + (y * stride);
-
-                for (int x = 0; x < imageData.Width; x++)
+                int firstx = 0, lastx = 0;
+                for (int x = 0; x < imageData.Width - TargetWMaxWidthMin; x++)
                 {
                     // Watch out for actual order (BGR)!
                     int bIndex = x * bytesPerPixel;
-                    int gIndex = bIndex + 1;
-                    int rIndex = bIndex + 2;
-                    if (row[rIndex] == searchedR && row[gIndex] == searchedG && row[bIndex] == searchedB)
+                    if (row[bIndex + 2] == searchedR && row[bIndex + 1] == searchedG && row[bIndex] == searchedB)
                     {
-                        if (count > 100 && (Math.Abs(centerX / count - x) > 50 || Math.Abs(centerY / count - y) > 50))
+                        row[bIndex + 2] = row[bIndex + 1] = row[bIndex] = 0;
+                        firstx = x;
+                        lastx = x + 1;
+                        while (lastx < imageData.Width)
                         {
+                            bIndex = lastx * bytesPerPixel;
+                            if (row[bIndex + 2] != searchedR || row[bIndex + 1] != searchedG && row[bIndex] != searchedB)
+                            {
+                                break;
+                            }
+                            row[bIndex + 2] = row[bIndex + 1] = row[bIndex] = 0;
+                            lastx++;
+                        }
+                        if (lastx - firstx < TargetWMaxWidthMin)
+                        {
+                            x = lastx;
                             continue;
                         }
-                        centerX += x;
-                        centerY += y;
-                        count++;
+                        else
+                        {
+                            break;
+                        }
                     }
                 }
+                var width = lastx - firstx;
+                if (width > TargetWMaxWidthMin && width > count)
+                {
+                    count = width;
+                    centerX = (lastx + firstx - 1) / 2;
+                    centerY = y;
+                }
+                if (count > TargetWMaxWidthMin && width < count)
+                {
+                    break;
+                }
             }
-            if (count > 0)
-            {
-                centerX /= count;
-                centerY /= count;
-            }
-
             image.UnlockBits(imageData);
         }
 
@@ -270,20 +126,14 @@ namespace PrecisionJump
                 {
                     // Watch out for actual order (BGR)!
                     int bIndex = x * bytesPerPixel;
-                    int gIndex = bIndex + 1;
-                    int rIndex = bIndex + 2;
-
-                    byte pixelR = row[rIndex];
-                    byte pixelG = row[gIndex];
-                    byte pixelB = row[bIndex];
-
-                    int diffR = pixelR - searchedR;
-                    int diffG = pixelG - searchedG;
-                    int diffB = pixelB - searchedB;
+                    int diffR = row[bIndex + 2] - searchedR;
+                    int diffG = row[bIndex + 1] - searchedG;
+                    int diffB = row[bIndex] - searchedB;
 
                     int distance = diffR * diffR + diffG * diffG + diffB * diffB;
                     if (distance < toleranceSquared)
                     {
+                        row[bIndex + 2] = row[bIndex + 1] = row[bIndex] = 0;
                         count++;
                         if (x < x1)
                         {
@@ -304,15 +154,17 @@ namespace PrecisionJump
             pcy = (y1 + y2) / 2.0f;
         }
 
-        static double MeasureDistance(string img)
+        static double MeasureDistance(string img, out bool direction)
         {
+            direction = false;
             using (var bitmap = new Bitmap(img))
             {
                 float tcx, tcy, pcx, pcy;
                 int count1, count2;
                 FindTargetWWithUnsafe(bitmap, TargetWR, TargetWG, TargetWB, out tcx, out tcy, out count1);
-                if (tcx < 10 || tcy < 10 || count1 < 600 || count1 > 900)
+                if (tcx < 10 || tcy < 10 || count1 < TargetWMaxWidthMin || count1 > 39)
                 {
+                    Debug.WriteLine($"-1:{img}");
                     return -1;
                 }
                 pcx = bitmap.Width + 40 - tcx;
@@ -321,18 +173,23 @@ namespace PrecisionJump
                 FindPersonAroundWithUnsafe(bitmap, PersonR, PersonG, PersonB, PersonTolerance, PersonSearchWith, PersonSearchHeight, ref pcx, ref pcy, out count2);
                 if (count2 < 2000)
                 {
-                    return -1;
+                    Debug.WriteLine($"-2:{img}");
+                    return -2;
                 }
 
-                //using (var g = Graphics.FromImage(bitmap))
-                //{
-                //    g.DrawString($"{count1} {count2}", new Font("Arial", 20), Brushes.Red, new PointF(100, 100));
-                //    g.FillEllipse(Brushes.Red, tcx - 2, tcy - 2, 5, 5);
-                //    g.FillEllipse(Brushes.Red, pcx - 2, pcy - 2, 5, 5);
-                //}
-                //bitmap.Save(Path.ChangeExtension(img, ".p.png"), ImageFormat.Png);
-                var dx = pcx - tcx;
-                var dy = pcy - tcy;
+                tcy += 2;
+                pcy += 7;
+
+                using (var g = Graphics.FromImage(bitmap))
+                {
+                    g.DrawString($"{count1} {count2}", new Font("Arial", 20), Brushes.Red, new PointF(100, 100));
+                    g.FillEllipse(Brushes.Red, tcx - 2, tcy - 2, 5, 5);
+                    g.FillEllipse(Brushes.Red, pcx - 2, pcy - 2, 5, 5);
+                }
+                bitmap.Save(Path.ChangeExtension(img, ".p.png"), ImageFormat.Png);
+                var dx = (pcx - tcx) / dpiX * Inch2Cm;
+                var dy = (pcy - tcy) / dpiY * Inch2Cm;
+                direction = dx < 0;
                 return Math.Sqrt(dx * dx + dy * dy);
             }
         }
@@ -344,16 +201,18 @@ namespace PrecisionJump
 
             var cmd1 = "shell /system/bin/screencap -p /sdcard/screenshot.png";
             var cmd2Tempate = "pull /sdcard/screenshot.png ";
-            var cmd3Tempate = "shell input swipe 250 250 300 300 ";
+            var cmd3Tempate = "shell input swipe 400 400 400 400 ";
             var cmdmodel = "shell getprop ro.product.model";
 
             var test = false;
-
+            var start = DateTime.Now;
             if (!test)
             {
                 var step = 0;
+                long[] tss = new long[5];
                 while (true)
                 {
+                    Console.Title = $" PrecisionJump {(DateTime.Now - start).TotalMinutes:F2}min";
                     var model = CmdAdb(cmdmodel);
                     if (string.IsNullOrWhiteSpace(model) || model.Contains("no devices"))
                     {
@@ -361,23 +220,45 @@ namespace PrecisionJump
                         break;
                     }
 
+                    tss[0] = DateTime.Now.Ticks;
                     CmdAdb(cmd1);
+
+                    tss[1] = DateTime.Now.Ticks;
                     var img = Path.Combine(screenshotspath, DateTime.Now.ToString("yyMMddHHmmssfff") + ".png");
                     CmdAdb($"{cmd2Tempate}{img}");
 
-                    var distance = MeasureDistance(img);
+                    tss[2] = DateTime.Now.Ticks;
+                    bool direction;
+                    var distance = MeasureDistance(img, out direction);
+
+                    tss[3] = DateTime.Now.Ticks;
                     if (distance > 0)
                     {
-                        var time = (int)(distance * InverseSpeed);
+                        step++;
+                        var time = direction ? (int)(distance * InvSpeedRight + OffsetRight) : (int)(distance * InvSpeedLeft + OffsetLeft);
                         CmdAdb($"{cmd3Tempate}{time}");
-                        Console.WriteLine($"Step:{step}\tDis:{distance}\tTime:{time}");
-                        // Thread.Sleep(time);
+                        tss[4] = DateTime.Now.Ticks;
+
+                        Console.WriteLine($"Step:{step}\tDis:{distance:F3}\tTime:{time}\tCMD1:{(tss[1] - tss[0]) / 1E7:F3}\tCMD2:{(tss[2] - tss[1]) / 1E7:F3}\tDIS:{(tss[3] - tss[2]) / 1E7:F3}\tCMD3:{(tss[4] - tss[3]) / 1E7:F3}");
+                        Thread.Sleep(1000);
                     }
                     else
                     {
-                        File.Delete(img);
                         Console.WriteLine($"Step:{step}\tCann't measure the distance");
-                        //Thread.Sleep(2000);
+                        if (distance == -1)
+                        {
+                            File.Delete(img);
+                            Thread.Sleep(2000);
+                        }
+                        else if (distance == -2)
+                        {
+                            File.Move(img, Path.ChangeExtension(img, ".2.png"));
+                            Thread.Sleep(1000);
+                        }
+                        else
+                        {
+                            // Thread.Sleep(100);
+                        }
                         continue;
                     }
                 }
@@ -387,8 +268,10 @@ namespace PrecisionJump
                 foreach (var file in Directory.EnumerateFiles(screenshotspath).ToList())
                 {
                     if (!file.EndsWith(".p.png"))
+                    //if (file.EndsWith("180113012303257.png"))
                     {
-                        if (MeasureDistance(file) < 0)
+                        bool direction;
+                        if (MeasureDistance(file, out direction) < 0)
                             Console.WriteLine(file);
                     }
                 }
